+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * [\#69](https://github.com/Finschia/wasmd/pull/69) refactor: refactor test cases for Params
 * [\#71](https://github.com/Finschia/wasmd/pull/71) add test cases in ContractsByCode
 * [\#82](https://github.com/Finschia/wasmd/pull/82) add test case to QueryInactiveContracts
+* [\#87](https://github.com/Finschia/wasmd/pull/87) add an integration test for TestExecuteContract 
 
 ### Bug Fixes
 * [\#62](https://github.com/Finschia/wasmd/pull/62) fill ContractHistory querier result's Updated field